Part of WOR-168: split test_watcher.py — remove duplicates, create finalize/promotion files (#312)
- Remove 42 duplicate test functions from test_watcher.py (helpers, subprocess,
  types, worktrees, services already covered in module-aligned files)
- Create tests/test_watcher_finalize.py (~508 LOC): all _finalize_worker test suites
  (PR failure, retry counting, set_state failures, usage-to-metrics, Sonar metrics,
  Sonar escalation, escalation policy flags)
- Create tests/test_watcher_promotion.py (~246 LOC): all _promote_waiting_tickets suites
- Append to test_watcher_types.py: test_write_and_remove_pid_file
- Append to test_watcher_worktrees.py: _cleanup_orphaned_worktrees tests
- test_watcher.py reduced from 1857 LOC → 397 LOC (Advisory tier)
- All imports point to sub-modules directly (no backward-compat watcher re-exports)
- 346 tests pass, coverage 84.78%, lint-imports contracts 4/4 kept
