+++ tests/test_watcher_types.py
@@ -4,7 +4,11 @@ from __future__ import annotations
 
 import os
 from pathlib import Path
+from unittest.mock import MagicMock
 
+import pytest
+
+from app.core.watcher import Watcher
 from app.core.watcher_types import is_watcher_running
 
 
@@ -26,3 +30,24 @@ def test_is_watcher_running_own_pid(tmp_path: Path) -> None:
     pid_file = tmp_path / "watcher.pid"
     pid_file.write_text(str(os.getpid()), encoding="utf-8")
     assert is_watcher_running(pid_file)
+
+
+# ---------------------------------------------------------------------------
+# Watcher PID file
+# ---------------------------------------------------------------------------
+
+
+def test_write_and_remove_pid_file(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    pid_file = tmp_path / ".claude/watcher.pid"
+    monkeypatch.setattr("app.core.watcher._PID_FILE", pid_file)
+
+    mock_linear = MagicMock()
+    watcher = Watcher(linear_client=mock_linear, repo_root=tmp_path)
+    watcher._write_pid_file()
+    assert pid_file.exists()
+    assert pid_file.read_text(encoding="utf-8") == str(os.getpid())
+
+    watcher._remove_pid_file()
+    assert not pid_file.exists()
